Uat

See merge request erp-group/cloud-erp-backend!767

+++ erp-server/erp-server-oms/src/main/java/com/erp/server/oms/sdk/sob2c/AliExpressSoB2cHandle.java
@@ -21,15 +21,19 @@ import com.erp.model.oms.enums.SoB2cErrorTypeEnum;
 import com.erp.model.wms.dto.AliexpressDeliveryDTO;
 import com.erp.model.wms.dto.AliexpressDeliveryDetailDTO;
 import com.erp.model.wms.dto.WarehouseMappingDTO;
+import com.erp.oms.aliexpress.dto.response.AliExpressDeliveryDetail;
+import com.erp.oms.aliexpress.service.AliExpressDliveryOrderService;
 import com.erp.rpc.dmp.feign.DmpTaskFeign;
 import com.erp.rpc.oms.feign.SkuMappingFeign;
 import com.erp.rpc.wms.feign.AliexpressDeliveryFeign;
 import com.erp.rpc.wms.feign.SoOutstockFeign;
 import com.erp.rpc.wms.feign.WarehouseMappingFeign;
 import com.erp.server.oms.service.*;
+import jnr.ffi.annotations.In;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -73,6 +77,8 @@ public class AliExpressSoB2cHandle implements ISoB2cHandleService {
     private SoB2cDetailService soB2cDetailService;
     @Resource
     private SoB2cService soB2cService;
+    @Autowired
+    private AliExpressDliveryOrderService aliExpressDliveryOrderService;
 
     @Override
     public Boolean handleRule(SoB2cEntity mainEntity) {
@@ -93,15 +99,10 @@ public class AliExpressSoB2cHandle implements ISoB2cHandleService {
     public Boolean handleSoOutStock(PlatformOrderDTO dto, SoB2cDTO.PullOrderResultDTO resultDTO, SoB2cEntity mainEntity) {
         //平台仓订单
         Boolean hasPlatformWarehouse = mainEntity.hasPlatformWarehouseOrder();
-//        Boolean isWarehouseEmpty = resultDTO.getIsWarehouseEmpty();
-        //已发货
-        String shipped = SoB2cBillStatusEnum.ENUM_SHIPPED.getCode();
-        String billStatus = mainEntity.getBillStatus();
-        boolean isShipped = shipped.equals(billStatus);
         try {
             // 平台仓生成销售出库单
             // 非平台仓由发货单生成销售出库单
-            if (isShipped && hasPlatformWarehouse) {
+            if (hasPlatformWarehouse) {
                 this.aliExpressDeliveryQuery(dto, resultDTO, mainEntity);
             }
             return true;
@@ -139,9 +140,10 @@ public class AliExpressSoB2cHandle implements ISoB2cHandleService {
             //根据仓库分组，同个仓库生成相同的销售出库单，销售出库单的sku和数量取速卖通返回的数据
             //根据平台sku查询Listing信息
             List<PlatformDeliveryDetailDTO> platformDeliveryDetailDTOList = dto.getDeliveryDetailDTOList();
-            List<String> platformSkuIdList = platformDeliveryDetailDTOList.stream().map(PlatformDeliveryDetailDTO::getPlatformSkuId).collect(Collectors.toList());
-            List<String> platformSpuList = platformDeliveryDetailDTOList.stream().map(PlatformDeliveryDetailDTO::getPlatformSpuNo).collect(Collectors.toList());
-            Map<String, List<ListingInfoWithSkuMappingDTO>> skuMappingMap = soB2cDetailService.mapListingByPlatformSkuId(platformSkuIdList, platformSpuList, mainEntity.getDictPlatform(), mainEntity.getShopId(), null, null);
+            platformDeliveryDetailDTOList = platformDeliveryDetailDTOList.stream().filter(v->StringUtils.isNotBlank(v.getPlatformWarehouseName())).collect(Collectors.toList());
+            if(CollectionUtils.isEmpty(platformDeliveryDetailDTOList)){
+                return;
+            }
 
             List<WarehouseMappingDTO.MappingViewDTO> mappingViewDTOS =  warehouseMappingFeign.listMappingViewByDictPlatform(mainEntity.getDictPlatform());
             //设置销售订单id，在后面新增销售出库单时用到
@@ -160,32 +162,26 @@ public class AliExpressSoB2cHandle implements ISoB2cHandleService {
                     soB2cErrorService.add(addError);
                     return;
                 }
+                List<String> platformSkuIdList = val.stream().map(PlatformDeliveryDetailDTO::getScItemId).distinct().collect(Collectors.toList());
+                List<SkuMappingDTO.WarehouseSkuDTO> warehouseSkuDTOList = skuMappingService.listByWarehouseAndPlatformSku(mappingViewDTO.getWarehouseId(),platformSkuIdList);
+                List<String> notMatchSkuNoList = new ArrayList<>();
                 for (PlatformDeliveryDetailDTO deliveryDetailDTO : val) {
-                    // 映射关系
-                    List<ListingInfoWithSkuMappingDTO> mappingDTOList = skuMappingMap.get(deliveryDetailDTO.getPlatformSkuId());
-                    // 检查和获取映射关系
-                    ListingInfoWithSkuMappingDTO mappingDTO = skuMappingService.checkAndMappingDTO(mappingDTOList, deliveryDetailDTO.getPlatformSpuNo(), mainEntity.getDictPlatform());
-
-                    if(Objects.isNull(mappingDTO) || StringUtils.isBlank(mappingDTO.getProductSkuId())){
-                        SoB2cErrorDTO.AddDTO addError = new SoB2cErrorDTO.AddDTO();
-                        addError.setType(SoB2cErrorTypeEnum.GENERATE_OUTSTOCK.getCode());
-                        addError.setParamJson("");
-                        addError.setReturnJson("");
-                        addError.setMainId(mainEntity.getId());
-                        addError.setMessage(StrUtil.format("自动生成销售出库单失败：订单未匹配Sku映射关系，平台产品ID:【{}】", deliveryDetailDTO.getPlatformSpuNo()));
-                        soB2cErrorService.add(addError);
-                        return;
+                    SkuMappingDTO.WarehouseSkuDTO warehouseSkuDTO = warehouseSkuDTOList.stream().filter(v->v.getPlatformSkuNo().equals(deliveryDetailDTO.getScItemId())).findFirst().orElse(new SkuMappingDTO.WarehouseSkuDTO());
+                    if(com.baomidou.mybatisplus.core.toolkit.StringUtils.isBlank(warehouseSkuDTO.getProductSkuId())){
+                        notMatchSkuNoList.add(deliveryDetailDTO.getScItemId());
+                        continue;
                     }
-                    deliveryDetailDTO.setSkuId(mappingDTO.getProductSkuId());
-                    deliveryDetailDTO.setSkuNo(mappingDTO.getProductSkuNo());
+                    deliveryDetailDTO.setSkuId(warehouseSkuDTO.getProductSkuId());
+                    deliveryDetailDTO.setSkuNo(warehouseSkuDTO.getProductSkuNo());
                     deliveryDetailDTO.setWarehouseId(mappingViewDTO.getWarehouseId());
                     deliveryDetailDTO.setWarehouseName(mappingViewDTO.getWarehouseName());
                     deliveryDetailDTO.setWarehouseOrgId(mappingViewDTO.getWarehouseOrgId());
                     deliveryDetailDTO.setWarehouseOrgName(mappingViewDTO.getWarehouseOrgName());
                 }
+                val = val.stream().filter(v->StringUtils.isNotBlank(v.getSkuId())).collect(Collectors.toList());
                 List<String> skuIdList = val.stream().map(PlatformDeliveryDetailDTO::getSkuId).distinct().collect(Collectors.toList());
                 PlatformGenerateSoOutstockDTO platformGenerateSoOutstockDTO = PlatformGenerateSoOutstockDTO.builder()
-                        .platformDeliveryDetailDTOList(platformDeliveryDetailDTOList)
+                        .platformDeliveryDetailDTOList(val)
                         .generateB2cDTO(soB2cService.getSoOutstockByIdAndWarehouseId(mainEntity.getId(),mappingViewDTO.getWarehouseId()))
                         .build();
                 //生成销售出库单
@@ -195,36 +191,21 @@ public class AliExpressSoB2cHandle implements ISoB2cHandleService {
                     soB2cDetailService.updateWarehouseByMapping(mappingViewDTO,mainEntity.getId(),skuIdList);
                     //生成速卖通发货单
                     addAliExpressDelivery(dto, mainEntity, logisticsDTOS, key,val);
+                }
+                if(CollectionUtils.isNotEmpty(notMatchSkuNoList)){
+                    SoB2cErrorDTO.AddDTO addError = new SoB2cErrorDTO.AddDTO();
+                    addError.setType(SoB2cErrorTypeEnum.GENERATE_OUTSTOCK.getCode());
+                    addError.setParamJson("");
+                    addError.setReturnJson("");
+                    addError.setMainId(mainEntity.getId());
+                    addError.setMessage(StrUtil.format("自动生成销售出库单失败：存在速卖通货品id未映射sku，货品id:【{}】", notMatchSkuNoList));
+                    soB2cErrorService.add(addError);
+                }else if (generateSoOutstockResult){
                     soB2cErrorService.deleteByCodeAndType(mainEntity.getCode(),SoB2cErrorTypeEnum.GENERATE_OUTSTOCK.getCode());
                     soB2cService.removeSignError(mainEntity.getId(),SoB2cErrorTypeEnum.GENERATE_OUTSTOCK.getCode());
                 }
             });
-        } else {
-            //如果仓库名称为空表示没找到速卖通发货单，记录异常订单，这里的WarehouseName是速卖通仓库名称
-            if (StringUtils.isBlank(warehouseName)) {
-                SoB2cErrorDTO.AddDTO addError = new SoB2cErrorDTO.AddDTO();
-                addError.setType(SoB2cErrorTypeEnum.GENERATE_OUTSTOCK.getCode());
-                addError.setParamJson("");
-                addError.setReturnJson("");
-                addError.setMainId(mainEntity.getId());
-                addError.setMessage("未查询到平台发货单或发货单未出库");
-                soB2cErrorService.add(addError);
-                return;
-            }
-
-            //是否匹配到仓库，是空表示未绑定速卖通仓库，记录异常订单
-            if (resultDTO.getIsWarehouseEmpty()) {
-                SoB2cErrorDTO.AddDTO addError = new SoB2cErrorDTO.AddDTO();
-                addError.setType(SoB2cErrorTypeEnum.GENERATE_OUTSTOCK.getCode());
-                addError.setParamJson("");
-                addError.setReturnJson("");
-                addError.setMainId(mainEntity.getId());
-                addError.setMessage(StrUtil.format("发货单仓库【{}】未匹配系统仓库", resultDTO.getWarehouseName()));
-                soB2cErrorService.add(addError);
-                return;
-            }
         }
-        return;
     }
 
     /**
